feat: Implemented api call within backend gameservice, added localview changer upon victory, downscale the video quality
