+++ backend/src/game/game.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
-import { GameService } from './game.service';
+import { GameClass } from './game.class';
 import { GameGateway } from './game.gateway';
+import { MatchModule } from 'src/match/match.module';
+import { GameService } from './game.service';
 
 @Module({
-  providers: [GameService, GameGateway],
+  imports: [MatchModule],
+  providers: [GameService, GameClass, GameGateway],
 })
 export class GameModule {}
